Abstracting helpers in rewrite-tutorials-link plugin
