+++ src/lib/remark-plugins/rewrite-tutorial-links/constants.ts
@@ -1,0 +1,3 @@
+const DEFAULT_CONTENT_TYPE = 'tutorials'
+
+export { DEFAULT_CONTENT_TYPE }
