+++ src/lib/remark-plugins/rewrite-tutorial-links/types.ts
@@ -7,4 +7,9 @@ type SplitLearnPath = [
 	string
 ]
 
-export type { SplitLearnPath }
+interface RewriteTutorialLinksPluginOptions {
+	contentType?: 'docs' | 'tutorials'
+	tutorialMap?: Record<string, string>
+}
+
+export type { RewriteTutorialLinksPluginOptions, SplitLearnPath }
